test both src and dist server. Update mocha test path
