+++ package.json
@@ -8,18 +8,17 @@
     "open:src": "babel-node tools/srcServer.js",
     "lint": "node_modules/.bin/esw webpack.config.* src tools",
     "lint:watch": "npm run lint -- --watch",
-    "test": "NODE_ENV='test' NODE_PATH=src mocha --reporter progress tools/testSetup.js \"src/**/*.spec.js\"",
-    "test:dev": "mocha --reporter spec tools/testSetup.js \"src/**/*.spec.js\"",
+    "test": "NODE_ENV='test' NODE_PATH=src mocha --reporter progress",
+    "test:dev": "mocha --reporter spec",
     "test:watch": "npm run test -- --watch",
-    "test:coverage": "rimraf coverage/ && istanbul cover _mocha tools/testSetup.js \"src/**/*.spec.js\"",
+    "test:coverage": "rimraf coverage/ && istanbul cover _mocha",
     "coveralls": "cat ./coverage/lcov.info | coveralls",
     "clean-dist": "npm run remove-dist && mkdir dist",
     "remove-dist": "node_modules/.bin/rimraf dist",
     "build:html": "babel-node tools/buildHtml.js",
     "prebuild": "npm-run-all clean-dist test build:html",
     "build": "babel-node tools/build.js NODE_ENV=production",
-    "postinstall": "npm run build",
-    "test:server": "node_modules/.bin/mocha --report lcovonly \"tools/tests/**/*.test.js\""
+    "postinstall": "npm run build"
   },
   "author": "Innocent Amadi",
   "license": "MIT",
