+++ tools/srcServer.js
@@ -38,3 +38,5 @@ app.listen(port, function (err) {
     console.log(`Server running on port: ${port}`)
   }
 });
+
+module.exports = app;
