+++ tools/tests/distServer.test.js
@@ -1,0 +1,4 @@
+const app = require('../distServer');
+
+const routeTest = require('./routeTest');
+routeTest(app);
