+++ tools/tests/srcServer.test.js
@@ -1,0 +1,4 @@
+const app = require('../srcServer');
+
+const routeTest = require('./routeTest');
+routeTest(app);
